Replace map[string]any with []slog.Attr
Now that log/slog is handling the logs, start using the provided
attributes. This brings the benefit of strong typing, among other
things. Changing from a map to a slice also retains the ordering of the
attributes. No more sorting the map by key. In the WithData method, make
a best effort attempt to retain the merging logic by enforcing unique
keys per logging entry and by letting new data have higher precedence
than previous data when items in either data list have the same key.

This is another breaking change to the exported API. Not a big deal
because the version of this package is sub v1.