test: Add tests for Infof, Errorf; check version key
Refactor the testLogg function to use subtests so that it's easier to
pinpoint errors. Also only check for the data key if there's data to be
expected. For package-level funcs like Infof and Errorf, there is no
extra data at the data key in logging entries. Bumps test coverage a
little bit too.