Replace Configure w/ Setup, add slog.Handler input
Breaking change b/c of removal of `Configure` function. The new
functionality allows you to change package-level variables rather than
forcing you to stay with the configuration once its been set for the
lifetime of the process.

Another breaking change is the modification of the New function.
Move the []slog.Attrs to the end so that it can leverage variadic args.

Add support for input slog.Handler. This gives an application more
control on where logs end up, how they're formatted and at what level to
log at.

The two functions, Setup and New have a similar signature. The use of
the []slog.Attr is different between the two. In the former, it is
versioning metadata, which would be present in subsequent logging events
at the "version" key. In the latter, the []slog.Attr is put at the
"data" key, and only for events emitted by the returned Emitter.

Consolidate examples