refactor: Change Emitter to a struct type
I think this was an unnecessary use of an interface. It was never
referenced by this library's internal functionality, and I don't think
there's a need for applications to implement the interface.

Technically, this is a breaking change because the type was changed but
the name stayed the same. Though it'd only break if an application
implemented the Emitter inteface, which would have been unnecessary IMO.

doc: Enhance documentation.

refactor: Move mergeAttrs closer to Emitter code.

test:
- Add regression test for dedupe behavior.
- Move some tests closer to Emitter type. Rename tests.
- Simplify tests for WithID, they were checking stuff in the data field
  for no good reason.