feat(internal): Add simple attribute handler
This is a slog.Handler implementation, mostly made to help me understand
how the slog package works. It could also be used for testing logging
behavior and outputs. Each piece of the handled message is captured in
the callback function. Its data can be inspected without having to think
about parsing the formatted data.