Simplify API again, decorate slog
OK, there are probably enough logger types for golang structured logging
libraries. Rather than add yet another one, delegate to the standard
library's log/slog package which will already be in golang versions 1.21
or higher.

The main thing I'm trying to achieve here is a particular data shape for
logging events. This package now offers a way to do that, adds a couple
of customization knobs, and gets out of the way.

Change default key names. The version key is now application_metadata
because this info could be from things other than version control
systems, such as build platform info, or whatever.

Remove the context-based API for getting and setting trace IDs. This was
unnecessary to put into this library. To add a tracing ID to a logger,
I think it'd be simpler to pass the tracing ID attribute directly.