fix: Ensure application_metadata key appears once
Log events from loggers created by the New function had a bug where the
application_metadata field appears twice. An explanation of how this
occurs:

The SetDefaults function calls the slog.SetDefaults function. When
making a logger with the New function, which used slog.Defaults, it
copies data from a slog.Handler that has had attrs and groups attached
to it. This leads to duplicate attributes, such as application_metadata.
Avoid using that handler by keeping a reference to the handler prior to
using it.