refactor AttrHandler
Change some state keeping mechanisms. Instead of a slice of groups or
attrs, use a linked list. Taking a cue from the slog handler guide and
from the github.com/jba/slog/withsupport package, add this type so that
calls to .WithAttrs, .WithGroup are running in constant time rather than
quadratic time.

Replace attribute merging. This was a lot of merging going on, and I'm
guessing added a lot of unnecessary CPU usage. The replacement
functionality keeps track of attributes with a map, and build upon
specific attributes. This requires finding a place for new attributes,
and creating space if there are some missing steps.

Add an attrBuilder type dedicated to creating the []slog.Attr for each
call to slog.Handler.Handle. All appends are now in one place. This type
also does attribute tracking, mentioned earlier in this commit msg.

Ensure that the handler code is doing all the work to pass the tests.
Before, there were some details, such as calling the .Resolve method,
which helped the tests from the testing/slogtest package pass.